Finished flame toggle, button interfaces, and activations.
Interval flame - got it to start and stop in game, on loop, made it a bp.
Toggle flame - got the toggle flame working, and the button interface to stop it. made it a bp
Confetti cannon. got the confetti cannon working, made it a bp.

only thing left for the flames is damaging player